separate classes for an easier refactoring
