+++ ui/src/main/AndroidManifest.xml
@@ -59,7 +59,7 @@
             android:name=".scanner.SelfScanningActivity"
             android:exported="false" />
         <activity
-            android:name=".scanner.CombinedScannerActivity"
+            android:name=".scanner.combined.CombinedScannerActivity"
             android:exported="false" />
         <activity
             android:name=".coupon.CouponDetailActivity"
