+++ ui/src/main/java/io/snabble/sdk/ui/scanner/combined/CombinedScannerActivity.kt
@@ -1,7 +1,8 @@
-package io.snabble.sdk.ui.scanner
+package io.snabble.sdk.ui.scanner.combined
 
 import io.snabble.sdk.ui.BaseFragmentActivity
 
 class CombinedScannerActivity : BaseFragmentActivity() {
+
     override fun onCreateFragment() = CombinedScannerFragment()
 }
